+++ core/navigation/src/main/java/com/goalpanzi/mission_mate/core/navigation/NavigationEventHandler.kt
@@ -3,7 +3,7 @@ package com.goalpanzi.mission_mate.core.navigation
 import kotlinx.coroutines.flow.SharedFlow
 
 interface NavigationEventHandler {
-    val routeToNavigate : SharedFlow<String>
+    val routeToNavigate : SharedFlow<RouteModel>
 
-    suspend fun triggerRouteToNavigate(route : String)
+    suspend fun triggerRouteToNavigate(route : RouteModel)
 }
